+++ src/data_sources/http/filter/request_body/mod.rs
@@ -1,4 +1,4 @@
-use async_graphql::{dynamic::ValueAccessor, Json};
+use async_graphql::Json;
 use bson::Document;
 use log::debug;
 
@@ -6,27 +6,24 @@ use crate::{data_sources::http::HttpDataSource, graphql::schema::ResolverType};
 
 impl HttpDataSource {
     pub fn create_body_filters(
-        input: &ValueAccessor<'_>,
+        mut input: Document,
         resolver_type: ResolverType,
     ) -> Option<Json<Document>> {
         debug!("Create Body Filters");
 
         match resolver_type {
             ResolverType::CreateOne => {
-                let document = input.deserialize::<Document>().unwrap();
-                debug!("Deserialized Document: {:#?}", document);
-
-                let json = Json::from(document);
+                let json = Json::from(input);
                 Some(json)
             }
             ResolverType::FindOne | ResolverType::FindMany => None,
             ResolverType::UpdateOne | ResolverType::UpdateMany => {
                 debug!("Update One Resolver Filters");
-                let mut document = input.deserialize::<Document>().unwrap();
-                document.remove("query");
-                let json = Json::from(document);
+                input.remove("query");
+                let json = Json::from(input);
                 Some(json)
             }
+            _ => panic!("Invalid resolver type"),
         }
     }
 }