+++ src/utils/document/mod.rs
@@ -1,0 +1,3 @@
+pub mod get_from_document;
+
+pub struct Document;